@@ -86,13 +86,13 @@ class User < ActiveRecord::Base
 
   before_save :encrypt_password, :unless => "password.blank?"
 
-  has_many :foster_dogs, :class_name => 'Dog', :foreign_key => 'foster_id'
-  has_many :current_foster_dogs, :class_name => 'Dog', :foreign_key => 'foster_id', :conditions => {:status => ['adoptable', 'adoption pending', 'on hold', 'coming soon', 'return pending']}
-  has_many :coordinated_dogs, :class_name => 'Dog', :foreign_key => 'coordinator_id', :conditions => {:status => ['adoptable', 'adopted', 'adoption pending', 'on hold', 'coming soon', 'return pending']}
+  has_many :foster_dogs, class_name: 'Dog', foreign_key: 'foster_id'
+  has_many :current_foster_dogs, -> { where(status: ['adoptable', 'adoption pending', 'on hold', 'coming soon', 'return pending']) }, class_name: 'Dog', foreign_key: 'foster_id'
+  has_many :coordinated_dogs, -> { where(status: ['adoptable', 'adopted', 'adoption pending', 'on hold', 'coming soon', 'return pending']) }, :class_name => 'Dog', foreign_key: 'coordinator_id'
   has_many :comments
   has_one :agreement, as: :attachable, class_name: 'Attachment' ,dependent: :destroy
   has_many :assignments, :class_name => 'Adopter', :foreign_key => 'assigned_to_user_id'
-  has_many :active_applications, :class_name => 'Adopter', :foreign_key => 'assigned_to_user_id', :conditions => {:status => ['new', 'pend response', 'workup', 'approved']}
+  has_many :active_applications, -> { where(status: ['new', 'pend response', 'workup', 'approved']) }, :class_name => 'Adopter', :foreign_key => 'assigned_to_user_id'
   belongs_to :mentor, :class_name => 'User', :foreign_key => 'mentor_id'
   has_many :mentees, :class_name => 'User', :foreign_key => 'mentor_id'
   accepts_nested_attributes_for :agreement
