@@ -117,8 +117,11 @@ Rails.application.configure do
     :url => ':s3_domain_url'
   }
 
-  config.middleware.use ExceptionNotifier,
-  sender_address: 'info@ophrescue.org',
-  exception_recipients: 'admin@ophrescue.org'
+  # Exception Notification
+  config.middleware.use ExceptionNotification::Rack,
+    email: {
+      sender_address: 'info@ophrescue.org',
+      exception_recipients: 'admin@ophrescue.org'
+    }
 
 end
