@@ -8,6 +8,7 @@ SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
 SimpleCov.start
 
 ENV['RAILS_ENV'] ||= 'test'
+require 'spec_helper'
 require File.expand_path('../../config/environment', __FILE__)
 require 'rspec/rails'
 
@@ -23,34 +24,36 @@ Capybara.javascript_driver = :poltergeist
 
 Dir[Rails.root.join('spec/support/**/*.rb')].sort.each { |file| require file }
 
+# Checks for pending migrations before tests are run.
+ActiveRecord::Migration.maintain_test_schema!
+
 RSpec.configure do |config|
-  config.run_all_when_everything_filtered = true
-  config.filter_run :focus
-
-  # Run specs in random order to surface order dependencies. If you find an
-  # order dependency and want to debug it, you can fix the order by providing
-  # the seed, which is printed after each run.
-  #     --seed 1234
-  config.order = 'random'
-
+  config.include ActiveSupport::Testing::TimeHelpers
   config.include FactoryGirl::Syntax::Methods
-  config.include JsonSpec::Helpers
+  config.include JsonSpec::Helpers, type: :request
+  config.include Rack::Test::Methods, type: :request
+
+  # Disable Rails transactional fixtures in favor of DatabaseCleaner
+  config.use_transactional_fixtures = false
 
   config.before(:suite) do
     DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.clean_with :truncation
+    DatabaseCleaner.clean_with(:truncation)
+
+    FactoryGirl.reload
   end
 
-  config.around(:each) do |example|
-    DatabaseCleaner.cleaning do
-      example.run
+  config.before(:each) do
+    DatabaseCleaner.start
   end
-end
 
   config.after(:each) do
     DatabaseCleaner.clean
   end
 
+  # Attempt to automatically mix in behaviours based on file location
+  # i.e. `get`, `post` in controller specs
+  config.infer_spec_type_from_file_location!
 end
 
 # Keep only the screenshots generated from the last failing test suite
@@ -61,36 +64,4 @@ Capybara::Screenshot.prune_strategy = :keep_last_run
 def send_keys_inputmask(location, keys)
   find(location).trigger('click')
   find(location).native.send_keys(keys)
-end
-
-class User
-  def chimp_subscribe
-    # no-op
-  end
-
-  def chimp_unsubscribe
-    # these happen in callbacks
-    # return true to not block
-    true
-  end
-end
-
-class Adopter
-  def chimp_subscribe
-    # these happen in callbacks
-    # return true to not block
-    true
-  end
-
-  def chimp_unsubscribe
-    # these happen in callbacks
-    # return true to not block
-    true
-  end
-
-  def chimp_check
-    # these happen in callbacks
-    # return true to not block
-    true
-  end
 end
